tests/test-requirements.txt
# API Tests
nose
ddt
jsonschema
opencafe
